requirements.txt
asgiref==3.4.1
Django==3.2.6
django-appconf==1.0.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-imagekit==4.0.2
django-modeltranslation==0.17.3
djangorestframework==3.12.4
pilkit==2.0
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
psycopg2
gunicorn
